Change CLI Installation
Update the CLI installation to use a @cli.installer("<lang>") approach
instead of using click. Also bake in a Context that builds up and
prepares all the information needed to perform the installation within
the decorator so that each installation operates the same and focuses
only on the specific logic required to install that CLI with that
language.